Add files via upload
This project is to implement the recommendation system. This system can help
company gain their profit because of promote their products. In this project, I need to
implement the recommendation system which can predict the score of movies. The
movies I predict have not seen by the user. we predict the movies score base on
another users’ movie score. We can predict through the result to recommend the user
movie they may like. The dataset is from MovieLen which provides 970 users score
in 9216 movies. I use k-nearest-neighbors (KNN) and Collaborative Filtering (CF) to
implement this system.